+++ test/tailwind_colors_test.dart
@@ -1,6 +1,4 @@
-import 'package:flutter_test/flutter_test.dart';
 
-import 'package:tailwind_colors/tailwind_colors.dart';
 
 void main() {